Added Ultimate Skill and updated damage system

+++ Assets/Scripts/Gameplay/Character/Weapon/Bullet.cs
@@ -8,9 +8,9 @@ public class Bullet : MonoBehaviour
 
     [SerializeField] private float _speed;
     [SerializeField] private Rigidbody _rb;
-    private int _damage;
+    private float _damage;
     
-    public void OnStart(int damage, int range)
+    public void OnStart(float damage, int range)
     {
         _damage = damage;
         _rb.velocity = transform.right * _speed;
@@ -21,7 +21,7 @@ public class Bullet : MonoBehaviour
     {
         if(other.TryGetComponent(out ZombieMelee zombie))
         {
-            zombie.ApplyDamage(_damage);
+            zombie.ApplyDamage((int)_damage);
             Destroy(gameObject);
         }
     }
